+++ src/archive/CartesianFrame.js
@@ -30,11 +30,11 @@ import defaultXYHTMLRule from './utils/defaultXYHTMLRule';
 import toAxes from './z-helper/toAxes';
 
 const naturalLanguageLineType = {
-  line: { items: 'index.js', chart: 'line chart' },
+  line: { items: 'Plot.js.js', chart: 'line chart' },
   area: { items: 'area', chart: 'area chart' },
-  cumulative: { items: 'index.js', chart: 'cumulative chart' },
-  'cumulative-reverse': { items: 'index.js', chart: 'cumulative chart' },
-  linepercent: { items: 'index.js', chart: 'line chart' },
+  cumulative: { items: 'Plot.js.js', chart: 'cumulative chart' },
+  'cumulative-reverse': { items: 'Plot.js.js', chart: 'cumulative chart' },
+  linepercent: { items: 'Plot.js.js', chart: 'line chart' },
   stackedarea: { items: 'stacked area', chart: 'stacked area chart' },
   'stackedarea-invert': { items: 'stacked area', chart: 'stacked area chart' },
   stackedpercent: { items: 'stacked area', chart: 'stacked area chart' },
@@ -46,7 +46,7 @@ const naturalLanguageLineType = {
   'bumparea-invert': { items: 'ranked area', chart: 'ranked area chart' },
   bumpline: { items: 'ranked line', chart: 'ranked line chart' },
   difference: {
-    items: 'index.js',
+    items: 'Plot.js.js',
     chart: 'difference chart'
   }
 };
@@ -191,7 +191,7 @@ const CartesianFrame = props => {
   };
 
   annotatedSettings.lineType.simpleLine =
-    annotatedSettings.lineType.type === 'index.js' &&
+    annotatedSettings.lineType.type === 'Plot.js.js' &&
     !annotatedSettings.lineType.y1 &&
     annotatedSettings.lineType.simpleLine !== false;
