+++ src/archive/utils/defaultXYSVGRule.js
@@ -175,7 +175,7 @@ const defaultXYSVGRule = ({
       xScale,
       yScale
     });
-  } else if (d.type === 'index.js') {
+  } else if (d.type === 'Plot.js.js') {
     return svgLineAnnotation({ d, i, screenCoordinates });
   } else if (d.type === 'area') {
     return svgAreaAnnotation({
