+++ src/frameUtils/index.js
@@ -1,7 +1,7 @@
 import getAdjustedPositionSize from './getAdjustedPositionSize';
 import toMarginGraphic from './toMarginGraphic';
 import generateFrameTitle from './generateFrameTitle';
-import getExtent from './getExtent';
+import getExtent from '../plots/BasePlot/getExtent';
 import toPipeline from './toPipeline';
 
 export {
