+++ src/plots/BasePlot/getExtent.js
