+++ src/plots/BasePlot/toRenderedAreas.js
