+++ src/plots/Contour/projection.js
@@ -87,7 +87,8 @@ const contouringProjection = ({
 
   return {
     projectedAreas,
-    projectedPoints
+    projectedPoints,
+    projectedLines: []
   };
 };
