+++ src/plots/Line/index.js
@@ -1,96 +1,11 @@
 import React from 'react';
-import {
-  object,
-  func,
-  array,
-  oneOfType,
-  bool,
-  number,
-  string,
-  node
-} from 'prop-types';
-import getExtent from '../../frameUtils/getExtent';
+import { object, func, array, oneOfType, bool, string, node } from 'prop-types';
 import lineProjection from './projection';
-import { stringToFn } from '../../archive/data/dataFunctions';
-import toRenderedLines from '../toRenderedLines';
-import toRenderedPoints from '../toRenderedPoints';
 
-const emptyObjectReturnFunction = () => ({});
-const emptyStringReturnFunction = () => '';
+import Plot from '../BasePlot';
 
 const Line = props => {
-  const {
-    data,
-    lineType,
-    lineStyle,
-    lineClass,
-    lineCustomMarks,
-    lineRenderMode,
-    pointStyle,
-    pointClass,
-    pointCustomMarks,
-    pointRenderMode,
-    useCanvas,
-    xAccessor,
-    yAccessor,
-    sAccessor,
-    xExtent,
-    yExtent,
-    showPoints,
-    frameXScale: xScale,
-    frameYScale: yScale,
-    size
-  } = props;
-
-  // extents
-  const { finalXExtent, finalYExtent } = getExtent({
-    data,
-    xAccessor,
-    yAccessor,
-    xExtent,
-    yExtent
-  });
-
-  // data projection
-  const { projectedAreas, projectedPoints, projectedLines } = lineProjection({
-    data,
-    finalXExtent,
-    finalYExtent,
-    xAccessor,
-    yAccessor,
-    sAccessor,
-    showPoints,
-    size
-  });
-
-  const { svgPipeline: lineSvg, canvasPipeline: lineCanvas } = toRenderedLines({
-    useCanvas,
-    xScale,
-    yScale,
-    styleFn: stringToFn(lineStyle, emptyObjectReturnFunction, true),
-    classFn: stringToFn(lineClass, emptyStringReturnFunction, true),
-    renderFn: stringToFn(lineRenderMode, undefined, true),
-    customMarks: lineCustomMarks,
-    data: projectedLines
-  });
-
-  const {
-    svgPipeline: pointsSvg,
-    canvasPipeline: pointsCanvas
-  } = toRenderedPoints({
-    useCanvas,
-    xScale,
-    yScale,
-    styleFn: stringToFn(pointStyle, emptyObjectReturnFunction, true),
-    classFn: stringToFn(pointClass, emptyStringReturnFunction, true),
-    renderFn: stringToFn(pointRenderMode, undefined, true),
-    customMarks: pointCustomMarks,
-    data: projectedPoints
-  });
-
-  const lineSvgPipeline = [...lineSvg, ...pointsSvg];
-  const lineCanvasPipeline = [...lineCanvas, ...pointsCanvas];
-  return lineSvgPipeline;
+  return <Plot projection={lineProjection} {...props} />;
 };
 
 Line.propTypes = {
